requirements.txt
replicate
python-dotenv
moviepy
requests
